change the way of controller are structured src/webservice/controllers/<controller>/index.ts → src/webservice/controllers/<controller>.ts
