+++ apps/io-services-cms-webapp/src/webservice/controllers/delete-service.ts
