+++ apps/io-services-cms-webapp/src/webservice/controllers/edit-service.ts
@@ -6,6 +6,7 @@ import {
 } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/azure_api_auth";
 import { RequiredParamMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/required_param";
 import { withRequestMiddlewares } from "@pagopa/io-functions-commons/dist/src/utils/request_middleware";
+import { RequiredBodyPayloadMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/required_body_payload";
 import {
   IResponseErrorConflict,
   IResponseErrorForbiddenNotAuthorized,
@@ -15,15 +16,22 @@ import {
   IResponseSuccessJson,
   IResponseSuccessNoContent,
   ResponseErrorInternal,
-  ResponseErrorNotFound,
   ResponseSuccessJson,
 } from "@pagopa/ts-commons/lib/responses";
 import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
-import * as O from "fp-ts/lib/Option";
-import * as TE from "fp-ts/lib/TaskEither";
-import { flow, pipe } from "fp-ts/lib/function";
-import { ServiceLifecycle as ServiceResponsePayload } from "../../../generated/api/ServiceLifecycle";
-import { itemToResponse } from "../../../utils/converters/service-lifecycle-converters";
+import * as RTE from "fp-ts/lib/ReaderTaskEither";
+import { pipe } from "fp-ts/lib/function";
+import { ServiceLifecycle as ServiceResponsePayload } from "../../generated/api/ServiceLifecycle";
+import { ServicePayload as ServiceRequestPayload } from "../../generated/api/ServicePayload";
+import {
+  itemToResponse,
+  payloadToItem,
+} from "../../utils/converters/service-lifecycle-converters";
+
+type Dependencies = {
+  // A store of ServicePublication objects
+  store: ReturnType<typeof stores.createCosmosStore<ServiceLifecycle.ItemType>>;
+};
 
 type HandlerResponseTypes =
   | IResponseSuccessJson<ServiceResponsePayload>
@@ -34,38 +42,26 @@ type HandlerResponseTypes =
   | IResponseErrorTooManyRequests
   | IResponseErrorInternal;
 
-type CreateServiceHandler = (
+type EditServiceHandler = (
   auth: IAzureApiAuthorization,
-  servicePayload: ServiceLifecycle.definitions.ServiceId
+  serviceId: ServiceLifecycle.definitions.ServiceId,
+  servicePayload: ServiceRequestPayload
 ) => Promise<HandlerResponseTypes>;
 
-type Dependencies = {
-  // A store od ServiceLifecycle objects
-  store: ReturnType<typeof stores.createCosmosStore<ServiceLifecycle.ItemType>>;
-};
-
-export const makeGetServiceLifecycleHandler =
-  ({ store }: Dependencies): CreateServiceHandler =>
-  (_auth, serviceId) =>
+export const makeEditServiceHandler =
+  ({ store }: Dependencies): EditServiceHandler =>
+  (_auth, serviceId, servicePayload) =>
     pipe(
-      serviceId,
-      store.fetch,
-      TE.bimap(
-        (err) => ResponseErrorInternal(err.message),
-        flow(
-          O.foldW(
-            () => ResponseErrorNotFound("Not found", `${serviceId} not found`),
-            (content) =>
-              ResponseSuccessJson<ServiceResponsePayload>(
-                itemToResponse(content)
-              )
-          )
-        )
-      ),
-      TE.toUnion
-    )();
+      ServiceLifecycle.apply("edit", serviceId, {
+        data: payloadToItem(serviceId, servicePayload),
+      }),
+      RTE.map(itemToResponse),
+      RTE.map(ResponseSuccessJson),
+      RTE.mapLeft((err) => ResponseErrorInternal(err.message)),
+      RTE.toUnion
+    )(store)();
 
-export const applyRequestMiddelwares = (handler: CreateServiceHandler) =>
+export const applyRequestMiddelwares = (handler: EditServiceHandler) =>
   pipe(
     handler,
     // TODO: implement ip filter
@@ -75,6 +71,8 @@ export const applyRequestMiddelwares = (handler: CreateServiceHandler) =>
       // only allow requests by users belonging to certain groups
       AzureApiAuthMiddleware(new Set([UserGroup.ApiServiceWrite])),
       // extract the service id from the path variables
-      RequiredParamMiddleware("serviceId", NonEmptyString)
+      RequiredParamMiddleware("serviceId", NonEmptyString),
+      // validate the reuqest body to be in the expected shape
+      RequiredBodyPayloadMiddleware(ServiceRequestPayload)
     )
   );
