+++ apps/io-services-cms-webapp/src/webservice/controllers/get-service-lifecycle.ts
@@ -6,7 +6,6 @@ import {
 } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/azure_api_auth";
 import { RequiredParamMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/required_param";
 import { withRequestMiddlewares } from "@pagopa/io-functions-commons/dist/src/utils/request_middleware";
-import { RequiredBodyPayloadMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/required_body_payload";
 import {
   IResponseErrorConflict,
   IResponseErrorForbiddenNotAuthorized,
@@ -16,22 +15,15 @@ import {
   IResponseSuccessJson,
   IResponseSuccessNoContent,
   ResponseErrorInternal,
+  ResponseErrorNotFound,
   ResponseSuccessJson,
 } from "@pagopa/ts-commons/lib/responses";
 import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
-import * as RTE from "fp-ts/lib/ReaderTaskEither";
-import { pipe } from "fp-ts/lib/function";
-import { ServiceLifecycle as ServiceResponsePayload } from "../../../generated/api/ServiceLifecycle";
-import { ServicePayload as ServiceRequestPayload } from "../../../generated/api/ServicePayload";
-import {
-  itemToResponse,
-  payloadToItem,
-} from "../../../utils/converters/service-lifecycle-converters";
-
-type Dependencies = {
-  // A store of ServicePublication objects
-  store: ReturnType<typeof stores.createCosmosStore<ServiceLifecycle.ItemType>>;
-};
+import * as O from "fp-ts/lib/Option";
+import * as TE from "fp-ts/lib/TaskEither";
+import { flow, pipe } from "fp-ts/lib/function";
+import { ServiceLifecycle as ServiceResponsePayload } from "../../generated/api/ServiceLifecycle";
+import { itemToResponse } from "../../utils/converters/service-lifecycle-converters";
 
 type HandlerResponseTypes =
   | IResponseSuccessJson<ServiceResponsePayload>
@@ -42,26 +34,38 @@ type HandlerResponseTypes =
   | IResponseErrorTooManyRequests
   | IResponseErrorInternal;
 
-type EditServiceHandler = (
+type CreateServiceHandler = (
   auth: IAzureApiAuthorization,
-  serviceId: ServiceLifecycle.definitions.ServiceId,
-  servicePayload: ServiceRequestPayload
+  servicePayload: ServiceLifecycle.definitions.ServiceId
 ) => Promise<HandlerResponseTypes>;
 
-export const makeEditServiceHandler =
-  ({ store }: Dependencies): EditServiceHandler =>
-  (_auth, serviceId, servicePayload) =>
-    pipe(
-      ServiceLifecycle.apply("edit", serviceId, {
-        data: payloadToItem(serviceId, servicePayload),
-      }),
-      RTE.map(itemToResponse),
-      RTE.map(ResponseSuccessJson),
-      RTE.mapLeft((err) => ResponseErrorInternal(err.message)),
-      RTE.toUnion
-    )(store)();
+type Dependencies = {
+  // A store od ServiceLifecycle objects
+  store: ReturnType<typeof stores.createCosmosStore<ServiceLifecycle.ItemType>>;
+};
 
-export const applyRequestMiddelwares = (handler: EditServiceHandler) =>
+export const makeGetServiceLifecycleHandler =
+  ({ store }: Dependencies): CreateServiceHandler =>
+  (_auth, serviceId) =>
+    pipe(
+      serviceId,
+      store.fetch,
+      TE.bimap(
+        (err) => ResponseErrorInternal(err.message),
+        flow(
+          O.foldW(
+            () => ResponseErrorNotFound("Not found", `${serviceId} not found`),
+            (content) =>
+              ResponseSuccessJson<ServiceResponsePayload>(
+                itemToResponse(content)
+              )
+          )
+        )
+      ),
+      TE.toUnion
+    )();
+
+export const applyRequestMiddelwares = (handler: CreateServiceHandler) =>
   pipe(
     handler,
     // TODO: implement ip filter
@@ -71,8 +75,6 @@ export const applyRequestMiddelwares = (handler: EditServiceHandler) =>
       // only allow requests by users belonging to certain groups
       AzureApiAuthMiddleware(new Set([UserGroup.ApiServiceWrite])),
       // extract the service id from the path variables
-      RequiredParamMiddleware("serviceId", NonEmptyString),
-      // validate the reuqest body to be in the expected shape
-      RequiredBodyPayloadMiddleware(ServiceRequestPayload)
+      RequiredParamMiddleware("serviceId", NonEmptyString)
     )
   );
